# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Gym Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /members:
    get:
      operationId: listMembers
      tags: [members]
      summary: List members with filters and pagination
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: search
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: planId
          in: query
          schema: { type: integer }
        - name: expiryWindow
          in: query
          schema: { type: integer }
        - name: sortBy
          in: query
          schema: { type: string }
        - name: sortOrder
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Paginated members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MembersPage"
    post:
      operationId: createMember
      tags: [members]
      summary: Create a new member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInput"
      responses:
        "201":
          description: Created member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get member by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    patch:
      operationId: updateMember
      tags: [members]
      summary: Update a member
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Archive (soft-delete) a member
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Archived
  /members/{id}/checkin:
    post:
      operationId: checkInMember
      tags: [members]
      summary: Check in a member
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckInBody"
      responses:
        "200":
          description: Check-in result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckInResult"
  /members/{id}/renew:
    post:
      operationId: renewMember
      tags: [members]
      summary: Renew a member's membership
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RenewBody"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}/freeze:
    post:
      operationId: freezeMember
      tags: [members]
      summary: Freeze a member's membership
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FreezeBody"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}/reactivate:
    post:
      operationId: reactivateMember
      tags: [members]
      summary: Reactivate a frozen or inactive member
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}/status:
    patch:
      operationId: setMemberStatus
      tags: [members]
      summary: Set member status (inactive, archived, active)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberStatusUpdate"
      responses:
        "200":
          description: Updated member
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
  /members/{id}/payments:
    get:
      operationId: getMemberPayments
      tags: [members]
      summary: Get payment history for a member
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Payment history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MemberPayment"
  /members/{id}/checkins:
    get:
      operationId: getMemberCheckins
      tags: [members]
      summary: Get check-in history for a member
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Check-in history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CheckInRecord"
...
  /attendance/member/{id}:
    get:
      operationId: getMemberAttendanceStats
      summary: Get attendance stats for a specific member
      tags: [checkins]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Member attendance statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberAttendanceStats"
...
components:
  schemas:
...
    Member:
      type: object
      properties:
        id: { type: integer }
        memberNumber: { type: string, nullable: true }
        name: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        emergencyContact: { type: string, nullable: true }
        gender: { type: string, nullable: true }
        planId: { type: integer, nullable: true }
        planName: { type: string, nullable: true }
        planPrice: { type: number, nullable: true }
        status: { type: string }
        joinDate: { type: string }
        startDate: { type: string, nullable: true }
        expiryDate: { type: string, nullable: true }
        lastCheckIn: { type: string, nullable: true }
        amountPaid: { type: number, nullable: true }
        discount: { type: number, nullable: true }
        balance: { type: number, nullable: true }
        currency: { type: string }
        frozenAt: { type: string, nullable: true }
        frozenUntil: { type: string, nullable: true }
        frozenDays: { type: integer, nullable: true }
        photoUrl: { type: string, nullable: true }
        fingerprintId: { type: string, nullable: true }
        qrCodeId: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, name, status, currency, joinDate, createdAt, updatedAt]
    MemberInput:
      type: object
      properties:
        name: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        emergencyContact: { type: string, nullable: true }
        gender: { type: string, nullable: true }
        joinDate: { type: string, nullable: true }
        planId: { type: integer, nullable: true }
        startDate: { type: string, nullable: true }
        expiryDate: { type: string, nullable: true }
        status: { type: string }
        amountPaid: { type: number }
        discount: { type: number }
        currency: { type: string, enum: [USD, CDF] }
        photoUrl: { type: string, nullable: true }
        fingerprintId: { type: string, nullable: true }
        qrCodeId: { type: string, nullable: true }
        notes: { type: string, nullable: true }
      required: [name, status, currency, amountPaid, discount]
    MemberUpdate:
      type: object
      properties:
        name: { type: string }
        email: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        address: { type: string, nullable: true }
        emergencyContact: { type: string, nullable: true }
        gender: { type: string, nullable: true }
        joinDate: { type: string, nullable: true }
        planId: { type: integer, nullable: true }
        startDate: { type: string, nullable: true }
        expiryDate: { type: string, nullable: true }
        status: { type: string, nullable: true }
        amountPaid: { type: number, nullable: true }
        discount: { type: number, nullable: true }
        currency: { type: string, nullable: true }
        photoUrl: { type: string, nullable: true }
        fingerprintId: { type: string, nullable: true }
        qrCodeId: { type: string, nullable: true }
        notes: { type: string, nullable: true }
    MembersPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Member"
        total: { type: integer }
        page: { type: integer }
        limit: { type: integer }
      required: [items, total, page, limit]
    CheckInBody:
      type: object
      properties:
        force:
          type: boolean
          description: Override duplicate same-day check-in
      required: []
    CheckInResult:
      type: object
      properties:
        success: { type: boolean }
        alreadyCheckedIn: { type: boolean }
        checkIn:
          $ref: "#/components/schemas/CheckInRecord"
      required: [success, alreadyCheckedIn]
    CheckInRecord:
      type: object
      properties:
        id: { type: integer }
        memberId: { type: integer }
        memberName: { type: string }
        memberNumber: { type: string, nullable: true }
        checkedInAt: { type: string }
        note: { type: string, nullable: true }
      required: [id, memberId, memberName, checkedInAt]
    RenewBody:
      type: object
      properties:
        planId: { type: integer }
        startDate: { type: string }
        expiryDate: { type: string }
        amountPaid: { type: number }
        discount: { type: number }
        currency: { type: string, enum: [USD, CDF] }
        notes: { type: string, nullable: true }
      required: [planId, startDate, expiryDate, amountPaid, discount, currency]
    FreezeBody:
      type: object
      properties:
        frozenAt: { type: string }
        frozenUntil: { type: string }
        reason: { type: string, nullable: true }
      required: [frozenAt, frozenUntil]
    MemberStatusUpdate:
      type: object
      properties:
        status:
          type: string
          enum: [active, inactive, archived]
      required: [status]
    MemberPayment:
      type: object
      properties:
        id: { type: integer }
        paymentNumber: { type: string, nullable: true }
        type: { type: string }
        amount: { type: number }
        discount: { type: number }
        currency: { type: string }
        planName: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, type, amount, discount, currency, createdAt]
...
    AttendanceMonthlyEntry:
      type: object
      properties:
        month: { type: string }
        count: { type: integer }
      required: [month, count]
...
    MemberAttendanceStats:
      type: object
      properties:
        totalCheckins: { type: integer }
        lastCheckIn: { type: string, nullable: true }
        thisMonthCheckins: { type: integer }
        avgPerMonth: { type: integer }
        monthlyHistory:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceMonthlyEntry"
        calendarDays:
          type: array
          items: { type: string }
      required: [totalCheckins, thisMonthCheckins, avgPerMonth, monthlyHistory, calendarDays]
...